Add block-network-request command
This provides a reasonable way to simulate a flakey network,
busted deploy, or ad blocker. It lets us actively block a
network request that the page makes.

    // to simulate JS that is enabled, but fails to load
    block-network-request: "*.js"

This is the simplest case, but is not the only useful one.
It can also block specific files, and it can stack:

    block-network-request: "*/search-index.js"
    block-network-request: "*/search.js"

Once a block list is added, there's no way to remove it.
Flakey network tests will probably be in separate test files,
so that shouldn't be a serious weakness.